Add files via upload
This Python script is like a handy notebook for keeping track of my network devices. I can start with a list of devices I have, and then add new ones as I find them. It also helps me sort them alphabetically, so they're easier to find later. I can even use it to see how many of a certain device I have, or where a specific device is on the list. On top of that, it lets me look at specific devices based on their position on the list, or grab a range of them at once. If I need to add a new device in a particular spot, I can do that too. And if things get messy, I can just clear the whole list and start fresh. There's even a bit of extra code at the end that shows how to loop through a list to print things, like all the even numbers from 0 to 8.